delete predictions api functions
